charger_jump-l4d2
Charger jump Improved

Here is a simple illustration:
survivors ----------------- you as a charger grabbing
survivors ---(shooting)---- a victim and tries to jump ---- victim is put down safely.
survivors ---------------- and died or seppuku.

Well, not anymore! By using this modified version, I added a convar along with a progress bar as a hint to notify players when will they be allowed to jump.

Here is the new illustration:
Same as above but --- victim (teleported with added velocity) then falls to his/her hilarious death.

Two Ways Of What Plugin Does:
It allows human-controlled charger to jump after 15 seconds only if he/she grabs a survivor and on ground. It doesn't work when charging for a long time then grabbing a survivor, not like that. Just charge then quickly grab a victim to be thrown in air.

When the delay is complete and said player jumps, both charger and victim will be registered that they are not on ground. If said charger seppuku, the victim will be teleported 3 times the value of charger_jump-l4d2_inertiavaultpower, and for all players, they will see the victim flying like he/she was super punched by a tank then eventually falls to death.
Another way to do that is if you seppuku or die in mid-air, you don't have to complete the delay, just press MIDDLE MOUSE BUTTON to seppuku and victim is automatically teleported.

Why it does that? Because you and victim are detected not on ground (that's what mid-air means).